# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  expedition.CreateExpeditionRequest:
    properties:
      endDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      games:
        items:
          $ref: '#/definitions/expedition.GameRequest'
        type: array
      imageUrls:
        items:
          type: string
        type: array
      isPublic:
        example: true
        type: boolean
      memo:
        example: 初めてのスタジアム訪問。とても楽しかった！
        type: string
      payments:
        items:
          $ref: '#/definitions/expedition.PaymentRequest'
        type: array
      sportId:
        example: 1
        type: integer
      stadiumId:
        example: 1
        type: integer
      startDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      title:
        example: 野球観戦の遠征記録
        type: string
      visitedFacilities:
        items:
          $ref: '#/definitions/expedition.VisitedFacilityRequest'
        type: array
    required:
    - endDate
    - games
    - imageUrls
    - isPublic
    - memo
    - payments
    - sportId
    - stadiumId
    - startDate
    - title
    - visitedFacilities
    type: object
  expedition.ExpeditionImageResponse:
    properties:
      fileId:
        example: file_1234567890
        type: string
      id:
        example: 1
        type: integer
      image:
        example: https://ik.imagekit.io/your_imagekit_id/image.jpg
        type: string
    type: object
  expedition.ExpeditionListResponse:
    properties:
      endDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      images:
        example:
        - https://ik.imagekit.io/your_imagekit_id/image.jpg
        items:
          type: string
        type: array
      likesCount:
        example: 10
        type: integer
      sportId:
        example: 1
        type: integer
      sportName:
        example: 野球
        type: string
      startDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      team1Name:
        example: ヤクルト
        type: string
      team2Name:
        example: ソフトバンク
        type: string
      title:
        example: 野球観戦の遠征記録
        type: string
      userIcon:
        example: https://ik.imagekit.io/your_imagekit_id/image.jpg
        type: string
      userId:
        example: 1
        type: integer
      userName:
        example: ユーザー名
        type: string
    type: object
  expedition.GameRequest:
    properties:
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      scores:
        items:
          $ref: '#/definitions/expedition.GameScoreRequest'
        type: array
      team1Id:
        example: 1
        type: integer
      team2Id:
        example: 2
        type: integer
    required:
    - date
    - scores
    - team1Id
    - team2Id
    type: object
  expedition.GameResponse:
    properties:
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      scores:
        items:
          $ref: '#/definitions/expedition.GameScoreResponse'
        type: array
      team1Id:
        example: 1
        type: integer
      team1Name:
        example: ヤクルト
        type: string
      team2Id:
        example: 2
        type: integer
      team2Name:
        example: ソフトバンク
        type: string
    type: object
  expedition.GameScoreRequest:
    properties:
      order:
        example: 1
        type: integer
      score:
        example: 1
        type: integer
      teamId:
        example: 1
        type: integer
    required:
    - order
    - score
    - teamId
    type: object
  expedition.GameScoreResponse:
    properties:
      id:
        example: 1
        type: integer
      order:
        example: 1
        type: integer
      score:
        example: 1
        type: integer
      teamId:
        example: 1
        type: integer
      teamName:
        example: ヤクルト
        type: string
    type: object
  expedition.GetExpeditionDetailResponse:
    properties:
      endDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      expeditionImages:
        items:
          $ref: '#/definitions/expedition.ExpeditionImageResponse'
        type: array
      games:
        items:
          $ref: '#/definitions/expedition.GameResponse'
        type: array
      id:
        example: 1
        type: integer
      isPublic:
        example: true
        type: boolean
      memo:
        example: 初めてのスタジアム訪問。とても楽しかった！
        type: string
      payments:
        items:
          $ref: '#/definitions/expedition.PaymentResponse'
        type: array
      sportId:
        example: 1
        type: integer
      sportName:
        example: 野球
        type: string
      stadiumId:
        example: 1
        type: integer
      stadiumName:
        example: 東京ドーム
        type: string
      startDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      title:
        example: 野球観戦の遠征記録
        type: string
      visitedFacilities:
        items:
          $ref: '#/definitions/expedition.VisitedFacilityResponse'
        type: array
    type: object
  expedition.PaymentRequest:
    properties:
      cost:
        example: 5000
        type: integer
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      title:
        example: チケット代
        type: string
    required:
    - cost
    - date
    - title
    type: object
  expedition.PaymentResponse:
    properties:
      cost:
        example: 5000
        type: integer
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      title:
        example: チケット代
        type: string
    type: object
  expedition.UpdateExpeditionImagesRequest:
    properties:
      add:
        items:
          type: string
        type: array
      delete:
        items:
          type: string
        type: array
    type: object
  expedition.UpdateExpeditionRequestBody:
    properties:
      endDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      games:
        $ref: '#/definitions/expedition.UpdateGamesRequest'
      images:
        $ref: '#/definitions/expedition.UpdateExpeditionImagesRequest'
      isPublic:
        example: true
        type: boolean
      memo:
        example: 初めてのスタジアム訪問。とても楽しかった！
        type: string
      payments:
        $ref: '#/definitions/expedition.UpdatePaymentsRequest'
      sportId:
        example: 1
        type: integer
      stadiumId:
        example: 1
        type: integer
      startDate:
        example: "2025-01-01T00:00:00Z"
        type: string
      title:
        example: 野球観戦の遠征記録
        type: string
      visitedFacilities:
        $ref: '#/definitions/expedition.UpdateVisitedFacilitiesRequest'
    required:
    - endDate
    - games
    - images
    - memo
    - payments
    - sportId
    - stadiumId
    - startDate
    - title
    - visitedFacilities
    type: object
  expedition.UpdateGameRequest:
    properties:
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      scores:
        $ref: '#/definitions/expedition.UpdateGameScoresRequest'
      team1Id:
        example: 1
        type: integer
      team2Id:
        example: 2
        type: integer
    required:
    - date
    - id
    - scores
    - team1Id
    - team2Id
    type: object
  expedition.UpdateGameScoreRequest:
    properties:
      id:
        example: 1
        type: integer
      order:
        example: 1
        type: integer
      score:
        example: 1
        type: integer
      teamId:
        example: 1
        type: integer
    required:
    - id
    - order
    - score
    - teamId
    type: object
  expedition.UpdateGameScoresRequest:
    properties:
      add:
        items:
          $ref: '#/definitions/expedition.GameScoreRequest'
        type: array
      delete:
        items:
          type: integer
        type: array
      update:
        items:
          $ref: '#/definitions/expedition.UpdateGameScoreRequest'
        type: array
    type: object
  expedition.UpdateGamesRequest:
    properties:
      add:
        items:
          $ref: '#/definitions/expedition.GameRequest'
        type: array
      delete:
        items:
          type: integer
        type: array
      update:
        items:
          $ref: '#/definitions/expedition.UpdateGameRequest'
        type: array
    type: object
  expedition.UpdatePaymentRequest:
    properties:
      cost:
        example: 5000
        type: integer
      date:
        example: "2025-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      title:
        example: チケット代
        type: string
    required:
    - cost
    - date
    - id
    - title
    type: object
  expedition.UpdatePaymentsRequest:
    properties:
      add:
        items:
          $ref: '#/definitions/expedition.PaymentRequest'
        type: array
      delete:
        items:
          type: integer
        type: array
      update:
        items:
          $ref: '#/definitions/expedition.UpdatePaymentRequest'
        type: array
    type: object
  expedition.UpdateVisitedFacilitiesRequest:
    properties:
      add:
        items:
          $ref: '#/definitions/expedition.VisitedFacilityRequest'
        type: array
      delete:
        items:
          type: integer
        type: array
      update:
        items:
          $ref: '#/definitions/expedition.UpdateVisitedFacilityRequest'
        type: array
    type: object
  expedition.UpdateVisitedFacilityRequest:
    properties:
      address:
        example: 東京都千代田区丸の内1-1-1
        type: string
      color:
        example: '#00FF00'
        type: string
      icon:
        example: train
        type: string
      id:
        example: 1
        type: integer
      latitude:
        example: 35.6812
        type: number
      longitude:
        example: 139.7671
        type: number
      name:
        example: 東京駅
        type: string
    required:
    - address
    - color
    - icon
    - id
    - latitude
    - longitude
    - name
    type: object
  expedition.VisitedFacilityRequest:
    properties:
      address:
        example: 東京都千代田区丸の内1-1-1
        type: string
      color:
        example: '#00FF00'
        type: string
      icon:
        example: train
        type: string
      latitude:
        example: 35.6812
        type: number
      longitude:
        example: 139.7671
        type: number
      name:
        example: 東京駅
        type: string
    required:
    - address
    - color
    - icon
    - latitude
    - longitude
    - name
    type: object
  expedition.VisitedFacilityResponse:
    properties:
      address:
        example: 東京都千代田区丸の内1-1-1
        type: string
      color:
        example: '#00FF00'
        type: string
      icon:
        example: train
        type: string
      id:
        example: 1
        type: integer
      latitude:
        example: 35.6812
        type: number
      longitude:
        example: 139.7671
        type: number
      name:
        example: 東京駅
        type: string
    type: object
  upload.UploadImagesResponse:
    properties:
      imageUrls:
        items:
          type: string
        type: array
    type: object
...
  utils.ApiResponse-array_expedition_ExpeditionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/expedition.ExpeditionListResponse'
        type: array
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  utils.ApiResponse-expedition_GetExpeditionDetailResponse:
    properties:
      data:
        $ref: '#/definitions/expedition.GetExpeditionDetailResponse'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
  utils.ApiResponse-upload_UploadImagesResponse:
    properties:
      data:
        $ref: '#/definitions/upload.UploadImagesResponse'
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  utils.ErrorBasicResponse:
    properties:
      message:
        example: エラーメッセージ
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessBasicResponse:
    properties:
      message:
        example: 成功しました！！
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: このapiは、ビジターゴーのAPIで、ユーザー、スタジアム、遠征記録、などについて扱います。
  title: ビジターゴーAPI
  version: "1.0"
paths:
...
  /api/admin/expedition/{expeditionId}:
    get:
      description: 遠征記録詳細情報を取得
      parameters:
      - description: expeditionId
        in: path
        name: expeditionId
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-expedition_GetExpeditionDetailResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: idから遠征記録詳細を取得
      tags:
      - expedition
  /api/admin/expedition/create:
    post:
      description: 遠征、出費、試合、訪れた施設の情報を保存する。
      parameters:
      - description: 遠征記録作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/expedition.CreateExpeditionRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録を作成
      tags:
      - expedition
  /api/admin/expedition/delete/{expeditionId}:
    delete:
      description: 遠征記録とそれに関連する全てのデータ（画像、いいね、支払い、試合情報など）を削除する
      parameters:
      - description: 遠征記録ID
        in: path
        name: expeditionId
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: 遠征記録が見つかりません
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録を削除
      tags:
      - expedition
  /api/admin/expedition/like/{expeditionId}:
    post:
      description: ユーザーが遠征記録にいいねを付ける
      parameters:
      - description: 遠征記録ID
        in: path
        name: expeditionId
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: 遠征記録が見つかりません
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録にいいねする
      tags:
      - expedition
  /api/admin/expedition/list:
    get:
      consumes:
      - application/json
      description: ページネーション付きで遠征記録一覧を取得します<br>teamIdとsportIdを指定すると、そのチーム、スポーツの遠征記録一覧を取得します。指定しなければ全ての遠征記録一覧を取得します
      parameters:
      - description: ページ番号
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: スポーツID
        in: query
        name: sportId
        type: integer
      - description: チームID
        in: query
        name: teamId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-array_expedition_ExpeditionListResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: 遠征記録が見つかりません
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録一覧を取得
      tags:
      - expedition
  /api/admin/expedition/unlike/{expeditionId}:
    delete:
      description: ユーザーが遠征記録のいいねを外す
      parameters:
      - description: 遠征記録ID
        in: path
        name: expeditionId
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: いいねが見つかりません
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録のいいねを外す
      tags:
      - expedition
  /api/admin/expedition/update/{expeditionId}:
    put:
      description: 遠征、出費、試合、訪れた施設の情報を更新する。<br>Payment, VisitedFacility, Game, GameScoreのdeleteにはidの配列ですが、ExpeditionImageのdeleteにはurlの配列をリクエストで渡してください
      parameters:
      - description: 遠征記録更新リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/expedition.UpdateExpeditionRequestBody'
      - description: 遠征記録ID
        in: path
        name: expeditionId
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.SuccessBasicResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "403":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: ユーザーが見つかりません
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      security:
      - BearerAuth: []
      summary: 遠征記録を更新
      tags:
      - expedition
...
  /api/admin/upload/images:
    post:
      consumes:
      - multipart/form-data
      description: 画像をアップロードし、URLを返します。<br>プロフィール、スタジアム、遠征など、格納フォルダを指定してください。<br>画像は1枚から10枚アップロードできるが、Swagger
        UIでは1つしか選択できません。<br>ファイルの拡張子は、[".jpg", ".jpeg", ".png"]だけを受け付けています。ファイルサイズは最大5MBを上限としています。
      parameters:
      - description: 格納フォルダ
        in: query
        name: folder
        required: true
        type: string
      - description: 画像ファイル
        in: formData
        name: images
        required: true
        type: file
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.ApiResponse-upload_UploadImagesResponse'
        "400":
          description: リクエストエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "404":
          description: not foundエラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/utils.ErrorBasicResponse'
      summary: 画像をクラウドストレージ(imagekit)にアップロード
      tags:
      - upload
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"